fix: make ep_uart/ep_can's octet-width read_size/segment_num casts explicit
Same MSVC C4244 pattern as the discovery.c fix: hdr.read_size_or_segment_num
is now the wire header's real 12-bit uint16_t field, but ep_uart.c's
out_read_size and both files' local segment_num were already,
independently, uint8_t -- pre-existing narrower types this PR's header
rework didn't touch. The narrowing (and its truncation ceiling for a
read_size/segment_num above 255) already existed before this PR;
making the cast explicit doesn't change behavior, it just satisfies
MSVC's /W4 the way GCC/Clang already silently did.

+++ src/ep_can.c
@@ -578,6 +578,12 @@ rcp_ep_can_errc_t rcp_ep_can_decode_frame_response_fragment(const uint8_t *b, si
     rcp_byte_bus_id_t             bus_id;
     uint8_t                       evt;
     uint8_t                       ms;
+    /* This local is this endpoint's own pre-existing octet-wide type
+     * (unchanged by the acf.c header rework); the wire field it's read
+     * from below is now the header's real 12-bit uint16_t field, so both
+     * assignments need an explicit narrowing cast for MSVC's /W4 -- same
+     * reasoning and same pre-existing (not newly introduced) truncation
+     * behavior as ep_uart.c's identical cast. */
     uint8_t                       segment_num;
     uint8_t                       transaction_num;
     bool                          timed;
@@ -594,7 +600,7 @@ rcp_ep_can_errc_t rcp_ep_can_decode_frame_response_fragment(const uint8_t *b, si
         bus_id          = gbb_hdr.info.byte_bus_id;
         evt             = gbb_hdr.info.evt;
         ms              = gbb_hdr.info.ms;
-        segment_num     = gbb_hdr.info.read_size_or_segment_num;
+        segment_num     = (uint8_t)gbb_hdr.info.read_size_or_segment_num;
         transaction_num = gbb_hdr.info.transaction_num;
         timed           = rcp_acf_gbb_is_timed(&gbb_hdr);
         timestamp       = timed ? gbb_hdr.message_timestamp : 0u;
@@ -606,7 +612,7 @@ rcp_ep_can_errc_t rcp_ep_can_decode_frame_response_fragment(const uint8_t *b, si
         bus_id          = abb_hdr.byte_bus_id;
         evt             = abb_hdr.evt;
         ms              = abb_hdr.ms;
-        segment_num     = abb_hdr.read_size_or_segment_num;
+        segment_num     = (uint8_t)abb_hdr.read_size_or_segment_num;
         transaction_num = abb_hdr.transaction_num;
         timed           = false;
         timestamp       = 0u;
